+++ src/main/java/org/cicirello/ds/BinaryHeap.java
@@ -73,9 +73,9 @@ import org.cicirello.util.Copyable;
  *     {@link #trimToSize}</li>
  * <li><b>O(m):</b> {@link #containsAll(Collection)}, {@link #createMaxHeap(Collection)}, 
  *     {@link #createMinHeap(Collection)}</li>
+ * <li><b>O(n + m):</b> {@link #retainAll(Collection)}</li>
  * <li><b>O(m lg n):</b> {@link #removeAll(Collection)}</li>
  * <li><b>O(m lg (n+m)):</b> {@link #addAll(Collection)}</li>
- * <li><b>O(n lg n):</b> {@link #retainAll(Collection)}</li>
  * </ul>
  *
  * @param <E> The type of object contained in the BinaryHeap.
@@ -492,24 +492,44 @@ public final class BinaryHeap<E> implements PriorityQueue<E>, Copyable<BinaryHea
 		return true;
 	}
 	
+	/**
+	 * {@inheritDoc}
+	 *
+	 * <p>The runtime of this method is O(n + m) where n is current size
+	 * of the heap and m is the size of the Collection c. In general this
+	 * is more efficient than calling remove repeatedly, unless you are
+	 * removing a relatively small number of elements, in which case you
+	 * should instead call {@link #remove(Object)} for each element you
+	 * want to remove.</p>
+	 */
 	@Override
 	public final boolean retainAll(Collection<?> c) {
-		HashSet<E> deleteThese = new HashSet<E>();
-		for (int i = 0; i < size; i++) {
-			deleteThese.add(buffer[i].element);
-		}
+		HashSet<Object> keepThese = new HashSet<Object>();
 		for (Object o : c) {
 			if (o instanceof PriorityQueueNode.Integer) {
 				PriorityQueueNode.Integer pair = (PriorityQueueNode.Integer)o;
-				deleteThese.remove(pair.element);
+				keepThese.add(pair.element);
 			} else {
-				deleteThese.remove(o);
+				keepThese.add(o);
 			}
 		}
 		boolean changed = false;
-		for (E e : deleteThese) {
-			remove(e); 
-			changed = true;
+		for (int i = size-1; i >= 0; i--) {
+			if (!keepThese.contains(buffer[i].element)) {
+				changed = true;
+				index.remove(buffer[i].element);
+				size--;
+				if (i == size) {
+					buffer[i] = null;
+				} else {
+					buffer[i] = buffer[size];
+					index.put(buffer[i].element, i);
+					buffer[size] = null;
+				}
+			}
+		}
+		if (changed) {
+			buildHeap();
 		}
 		return changed;
 	}