+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-05-26
+## [Unreleased] - 2022-06-05
 
 ### Added
 
 ### Changed
+* Improved implementation of BinaryHeap.retainAll(Collection) to an O(m + n) runtime.
+* Improved implementation of BinaryHeapDouble.retainAll(Collection) to an O(m + n) runtime.
 
 ### Deprecated